Time: 11 ms (65.17%), Space: 11.6 MB (88.57%) - LeetHub

+++ 61-rotate-list/61-rotate-list.cpp
@@ -21,7 +21,7 @@ public:
             return head;
         }
         else{
-            long long int limit=1;
+            int limit=1;
             ListNode* lim= head;
             while(lim->next!=NULL)
             {
